requirements.txt
accelerate==0.19.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
appnope==0.1.3
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bert-score==0.3.7
blinker==1.6.2
blis==0.4.1
braceexpand==0.1.7
cachetools==5.3.0
catalogue==1.0.2
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
contexttimer==0.3.3
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
decord==0.6.0
distlib==0.3.6
einops==0.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
entrypoints==0.4
fairscale==0.4.4
filelock==3.12.0
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.14.1
identify==2.5.24
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.10
ipython==7.34.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kaggle==1.5.13
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
murmurhash==1.0.9
networkx==2.6.3
nodeenv==1.8.0
numpy==1.19.2
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.1
pandas==1.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==3.5.1
plotly==5.14.1
portalocker==2.7.0
pre-commit==2.21.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyarrow==12.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
rich==13.3.5
sacremoses==0.0.53
salesforce-lavis==1.0.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
spacy==2.3.2
srsly==1.0.6
streamlit==1.22.0
tenacity==8.2.2
text-unidecode==1.3
thinc==7.4.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
typing_extensions==4.5.0
tzlocal==5.0
urllib3==2.0.2
validators==0.20.0
virtualenv==20.23.0
wasabi==0.10.1
wcwidth==0.2.6
webdataset==0.2.48
zipp==3.15.0
